backend/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
japanize-matplotlib==1.1.3
pyjpholiday==0.4.1
